Fix off-by-one errors and remove Save button.
Fix the off-by-one errors which were occurring and causing the
"next" year to switch back to the previous year when
creating the commit schedule. This wasn't noticed during initial
development because the commit schedule wasn't previously generated
at the end of a year.

Remove the Save button which is currently in the backlog of features
to be developed.